Backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
beautifulsoup4==4.12.2
CacheControl==0.14.4
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cryptography==48.0.0
dnspython==2.8.0
fastapi==0.104.1
firebase_admin==7.4.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.195.0
google-auth==2.50.0
google-auth-httplib2==0.3.1
google-cloud-core==2.5.1
google-cloud-firestore==2.26.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-generativeai==0.8.6
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
greenlet==3.5.0
grpcio==1.80.0
grpcio-status>=1.71,<1.75
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.13
motor==3.3.2
msgpack==1.1.2
numpy==2.4.4
playwright==1.59.0
proto-plus==1.27.2
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pyee==13.0.1
PyGithub==2.9.1
PyJWT==2.12.1
pymongo==4.6.1
PyNaCl==1.6.2
pyparsing==3.3.2
pyrate-limiter==4.1.0
python-dotenv==1.2.2
PyYAML==6.0.3
rank-bm25==0.2.2
requests==2.33.1
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.27.0
tqdm==4.67.3
tree-sitter==0.25.2
tree-sitter-java==0.23.5
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==12.0
